console logs and comments removed

+++ models/news.models.js
@@ -9,15 +9,19 @@ exports.fetchTopics = () => {
 };
 
 exports.fetchArticlesById = (article_id) => {
-    if ({ rows } < 1) { return Promise.reject({ status: 404, msg: 'Path not found' }); 
+    return db
+    .query(`SELECT * FROM articles WHERE article_id = $1;`, [article_id])
+    .then(({ rows }) => {
+        if ( rows.length < 1 ) {
+        return Promise.reject({
+            status: 404,
+            msg: `Resource not found`,
+            });
     }
     else {
-    return db
-    .query('SELECT * FROM articles WHERE article_id = $1;', [article_id])
-    .then(({ rows }) => {
         return rows[0];
-    });
-    };
+    }
+    })
 };
 
 exports.updateArticleById = (article_id, updatedArticle) => {
@@ -50,8 +54,12 @@ exports.fetchArticles = () => {
 };
 
 exports.deleteCommentById = (comment_id) => {
+    if ({ rows } < 1) { return Promise.reject({ status: 404, msg: 'Path not found' }); 
+    }
+    else {
     return db
     .query('DELETE FROM comments WHERE comment_id = $1 RETURNING *;', [comment_id])
+    }
 }
